refactor: 기관 추가 table-item 분리
